revert back to 8 bit tq value setting minimum send and minimum recv to "1" setting hop penalty to "10"
git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@983 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ ring_buffer.h
@@ -24,5 +24,5 @@
 
 
 
-void ring_buffer_set(uint16_t tq_recv[], uint8_t *tq_index, uint16_t value);
-uint16_t ring_buffer_avg(uint16_t tq_recv[]);
+void ring_buffer_set(uint8_t tq_recv[], uint8_t *tq_index, uint8_t value);
+uint8_t ring_buffer_avg(uint8_t tq_recv[]);
